게시글 읽기 API에서 잘못된 게시글의 조회수가 증가하는 버그 수정, 게시글의 댓글 수 추가

+++ server/src/posts/commands/handlers/delete-post.handler.ts
@@ -5,6 +5,7 @@ import { UsersRepository } from "../../../users/users.repository";
 import { TagChangedEvent } from "../../../tags/events/tag-changed.event";
 import { PostImageChangedEvent } from "../../../upload/event/post-image-changed.event";
 import { PostDeletedEvent } from "../../events/post-deleted.event";
+import { NotFoundException } from "@nestjs/common";
 
 @CommandHandler(DeletePostCommand)
 export class DeletePostHandler implements ICommandHandler<DeletePostCommand> {
@@ -20,6 +21,7 @@ export class DeletePostHandler implements ICommandHandler<DeletePostCommand> {
       requestUserNickname,
     );
     const post = await this.postRepository.findByPostId(postIdentifier);
+    if (!post) throw new NotFoundException("잘못된 게시글");
     post.verifyOwner(requestUser);
     await this.postRepository.remove(post);
